backend/requirements.txt
# Django et ses dépendances
asgiref==3.7.2
Django==4.2.9
sqlparse==0.4.4

# REST Framework et extensions
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-nested-routers==0.93.4

# Base de données
psycopg2-binary==2.9.9

# AWS et stockage
boto3==1.34.69
botocore==1.34.69
django-storages==1.14.2
s3transfer==0.10.0

# Utilitaires
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
urllib3==2.0.7

# GenIA dependencies
requests==2.32.3
PyPDF2==3.0.1